// src/types/miscTypes.ts
export interface BadgeListType {
    ActiveDeveloper: string;
    BugHunterLevel1: string;
    BugHunterLevel2: string;
    PremiumEarlySupporter: string;
    Partner: string;
    Staff: string;
    HypeSquadOnlineHouse1: string;
    HypeSquadOnlineHouse2: string;
    HypeSquadOnlineHouse3: string;
    Hypesquad: string;
    CertifiedModerator: string;
    VerifiedDeveloper: string;
}